scraper/requirements.txt
requests>=2.31.0
pandas>=2.0.0
google-cloud-bigquery>=3.10.0
google-auth>=2.22.0
pyarrow>=14.0.0
db-dtypes>=1.1.1
